+++ R/use_calibrations_bladj.R
@@ -1,19 +1,19 @@
-# calibrations <- get_all_calibrations(phyloall)
-# phy <- use_calibrations_bladj(phy, calibrations, type = "median")
-
-# calibrations <- get_all_calibrations(cetaceae_phyloall)
-# phy <- cetaceae_phyloall[[2]]
-# plot(use_all_calibrations_bladj(phy,calibrations, use = "Mean"))
-
 #' Use calibrations to date a topology with the BLADJ algorithm.
 #'
-#' @description The function prepares the input for BLADJ and calls [make_bladj_tree()]
+#' @description The function `use_calibrations_bladj` prepares the input for BLADJ
+#' and calls [make_bladj_tree()].
 #' @param phy A `phylo` object with or without branch lengths.
 #' @param calibrations A `data.frame` of secondary calibrations for any pair of taxon
 #' names in `phy`, usually obtained with [get_all_calibrations()].
 #' @param type The type of age to use as calibration. Options are "median", "mean", "min", or "max".
-#' @param root_age Not implemented yet. Numeric specifying the age of the root. If there are no calibrations for it. If NULL or not numeric, the maximum calibration plus a unit of the mean differences will be used as root calibration. If there is only one internal calibration, the root age will be set to 10% more than the age of the calibration.
-#' @param match_calibrations Boolean, default to TRUE. It will run match_all_calibrations function. Set to FALSE if your calibrations have already been matched.
+#' @param root_age Not implemented yet. Numeric specifying the age of the root.
+#' If there are no calibrations for it. If NULL or not numeric, the maximum
+#' calibration plus a unit of the mean differences will be used as root calibration.
+#' If there is only one internal calibration, the root age will be set to 10% more
+#' than the age of the calibration.
+#' @param match_calibrations Boolean, default to TRUE. It will run
+#' [match_all_calibrations()] function. Set to FALSE if your ages have
+#' already been matched.
 #' @return A `phylo` object with branch lengths proportional to time.
 #' @details
 #' The BLADJ algorithm is part of the Phylocom software, presented in Webb et al.
@@ -24,69 +24,71 @@
 #' Bioinformatics, 24(18), \doi{10.1093/bioinformatics/btn358}.
 
 #' @export
-use_calibrations_bladj <- function(phy,
+use_calibrations_bladj <- function(phy = NULL,
                                    calibrations,
                                    type = "median",
                                    root_age = NULL,
-                                   match_calibrations = TRUE) {
+                                   match_calibrations = TRUE,
+                                   congruified = FALSE) {
   #
-  message("... Using secondary calibrations with BLADJ.")
-  if (!inherits(phy, "phylo")) {
-    stop("'phy' is not a phylo object.")
+  message("... Using ages as secondary calibrations with BLADJ.")
+  if (inherits(calibrations, "matchedCalibrations")) {
+      phy <- calibrations$phy
+      calibs <- calibrations$matched_calibrations
+    stop("Not implemented yet, apologies.")
   }
-  if (!inherits(calibrations, "data.frame")) {
-    stop("'calibrations' is not a data.frame, dating with BLADJ is not possible.\n\t Provide a set of calibrations for 'phy', hint: see get_all_calibrations function.")
+  if (inherits(calibrations, "congruifiedCalibrations")) {
+    phy <- attributes(calibrations)$phy
+
   }
   type <- match.arg(tolower(type), c("mean", "min", "max", "median"))
-  if (match_calibrations) {
-    calibs <- match_all_calibrations(phy, calibrations)
-  } else {
-    calibs <- calibrations
-    # calibs <- all_calibs_93_matched
-  }
+
   if (nrow(calibs$matched_calibrations) < 1) {
     stop("Nodes in 'calibrations' (determined by taxon pairs) do not match any nodes
-						in 'phy'.\n\t Dating analysis is not possible with this set of calibrations.")
+            in 'phy'.\n\t Dating analysis is not possible with this set of calibrations.")
   }
+  age_distributions <- calibs$phy$calibration_distribution
+  node_names <- calibs$matched_calibrations$NodeNames
+
+
   if ("mean" %in% type) {
-    node_ages <- sapply(calibs$phy$calibration_distribution, mean)
+    node_ages <- sapply(age_distributions, mean)
     # length(node_ages)
   }
   if ("min" %in% type) {
-    node_ages <- sapply(calibs$phy$calibration_distribution, min)
+    node_ages <- sapply(age_distributions, min)
   }
   if ("max" %in% type) {
-    node_ages <- sapply(calibs$phy$calibration_distribution, max)
+    node_ages <- sapply(age_distributions, max)
   }
   if ("median" %in% type) {
-    node_ages <- sapply(calibs$phy$calibration_distribution, stats::median)
+    node_ages <- sapply(age_distributions, stats::median)
   }
   # check that the root node has a calibration
   # otherwise, add one
-  # bladj does not run if there is no calibration for the root
-  # bladj will run if calibrations are in conflict
-  node_names <- calibs$matched_calibrations$NodeNames
-  # length(node_names)
-  if (!"n1" %in% calibs$matched_calibrations$NodeNames) {
+  # bladj stops if there is no calibration for the root
+  if (!"n1" %in% node_names) {
     if (length(node_ages) > 1) {
+      # if there is only one calibration the line below gives give NaN
       root_age <- max(node_ages) + mean(abs(diff(sort(node_ages))))
     } else {
-      # if there is only one calibration the line above will give NaN
+      # case when there is only one calibration:
       root_age <- 1.1 * max(node_ages)
     }
     node_ages <- c(root_age, node_ages)
     node_names <- c("n1", node_names)
   }
-  new_phy <- make_bladj_tree(
+  # bladj freezes if calibrations are in conflict
+  chronogram <- make_bladj_tree(
     tree = calibs$phy, nodeages = node_ages,
     nodenames = node_names
   )
   # TODO: something about these extra list elements, set up as attributes??
-  new_phy$dating_method <- "bladj"
-  new_phy$calibration_distribution <- calibs$phy$calibration_distribution
-  new_phy$calibrations <- calibs$matched_calibrations
-  new_phy$used_calibrations <- stats::setNames(node_ages, node_names)
-  return(new_phy)
+  chronogram$dating_method <- "bladj"
+  chronogram$calibration_distribution <- ages_distribution
+  chronogram$calibrations <- calibs$matched_calibrations
+  chronogram$used_calibrations <- stats::setNames(node_ages, node_names)
+  return(chronogram)
 }