+++ NEWS.md
@@ -1,3 +1,10 @@
+
+TODO:
+  - add congruification step to `datelife_use` functions
+  - `use_calibrations_bladj` should take an output of `congruify_andd_mrca`
+
+DONE:
+
 # datelife 0.6.2
   - data: updated `opentree_chronograms` object. It now has 253 chronograms from Open Tree of Life
   - examples: set to not run most examples
@@ -6,7 +13,8 @@
     - bug fix on `use_calibrations_bladj` that used element $present_calibrations instead of $matched_calibrations
     - `get_otol_chronogram` is another name to call `get_opentree_chronograms`
     - added taxonomic source options argument to `make_datelife_query`. You can choose from OTT, NCBI, IRMNG and GBIF.
-  - added function method `congruify and mrca`
+    - function `congruify_and_mrca`: output has congruified topology with nodelables as attribute.
+  - added function method `congruify_and_mrca`
 
 # datelife 0.6.1
   - Functions: